modification + toggle fix
