+++ src/parser.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"errors"
-	"fmt"
 	"strconv"
 	"time"
 	"umutsevdi/td/todo"
@@ -16,34 +15,33 @@ func (e Error) Error() string { return string(e) }
 
 const argError = Error("Error in argument")
 
-func ParseArgs(args cli.Args) (todo.Todo, error) {
+func FindAndReplace(args cli.Args, t *todo.Todo) error {
 	// description case
-	var t todo.Todo
 	if args.Len() == 2 {
 		t.Desc = args.Get(1)
-		return t, nil
+		return nil
 	}
 	// key value case
 	for i := 2; i < args.Len(); i += 2 {
-		err := parseEntry(&t, args.Get(i), args.Get(i-1))
+		err := parseEntry(t, args.Get(i-1), args.Get(i))
 		if err != nil {
-			return todo.Todo{}, err
+			return err
 		}
 	}
-	return t, nil
+	return nil
 }
 
 func parseEntry(t *todo.Todo, key, value string) error {
 	switch key {
-	case "desc", "d":
+	case "--desc", "-d":
 		t.Desc = value
-	case "date", "D":
+	case "--date", "-D":
 		d, err := parseDate(value)
 		if err != nil {
 			return err
 		}
 		t.Deadline = d
-	case "period", "p":
+	case "--period", "-p":
 		v, err := strconv.Atoi(value)
 		if err != nil {
 			return err
@@ -66,18 +64,18 @@ func parsePeriod(v string) (int, error) {
 }
 
 func parseDate(v string) (time.Time, error) {
+	if len(v) == 0 {
+		return time.Time{}, nil
+	}
 	d, err := time.Parse("02/01/2006 15:04", v)
 	if err != nil {
-        fmt.Println(err)
 		d, err = time.Parse("02/01/2006", v)
 		if err != nil {
-            fmt.Println(err)
 			d, err = time.Parse("15:04", v)
 			if err != nil {
-                fmt.Println(err)
 				return time.Time{}, errors.New("invalid date format")
 			}
-			d.AddDate(time.Now().Day(), int(time.Now().Month()), time.Now().Year())
+			d = d.AddDate(time.Now().Year(), int(time.Now().Month())-1, time.Now().Day()-1)
 		}
 	}
 	return d, nil
